Move DB schema to separate module
