+++ economy/db.py
@@ -1,13 +1,21 @@
 from contextlib import contextmanager
-from sqlalchemy import create_engine, Column, String, Integer, Float
-from sqlalchemy.orm import sessionmaker, declarative_base
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
 
 # Database path is configured via the shared config module
 from config import DB_PATH
 
+from .schema import (
+    Base,
+    GoodsTable,
+    JobsTable,
+    JobInput,
+    JobOutput,
+    JobTool,
+)
+
 engine = create_engine(f"sqlite:///{DB_PATH}")
 SessionLocal = sessionmaker(bind=engine)
-Base = declarative_base()
 
 
 def get_session():
@@ -23,43 +31,6 @@ def session_scope():
         yield session
     finally:
         session.close()
-
-
-class GoodsTable(Base):
-    __tablename__ = "goods"
-    name = Column(String, primary_key=True)
-    size = Column(Float)
-
-
-class JobsTable(Base):
-    __tablename__ = "jobs"
-    name = Column(String, primary_key=True)
-    job_limit = Column(Integer)
-
-
-class JobInput(Base):
-    __tablename__ = "job_inputs"
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    job = Column(String)
-    good = Column(String)
-    qty = Column(Integer)
-
-
-class JobOutput(Base):
-    __tablename__ = "job_outputs"
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    job = Column(String)
-    good = Column(String)
-    qty = Column(Integer)
-
-
-class JobTool(Base):
-    __tablename__ = "job_tools"
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    job = Column(String)
-    tool = Column(String)
-    qty = Column(Integer)
-    break_chance = Column(Float)
 
 
 def rebuild_database():
